refactor: use-folder로 코드 통합
