--- src/pages/archive/hooks/use-post-folders.ts
@@ -1,14 +1,0 @@
-import { FolderData, postFolders } from "@apis/folder/folder";
-import { FOLDER_QUERY_KEY } from "@apis/folder/folder-queries";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-
-export const usePostFolders = () => {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: (body: FolderData) => postFolders(body),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: FOLDER_QUERY_KEY.ALL() });
-    },
-  });
-};
